Removed ability to create multiple posts, added ability to send freeform request data
I realized that based on the documentation and how the api worked, the ability to create multiple posts must be either an undocumented feauter, or a bug. Either way as QA, I'd file a bug on it, because it shouldn't accept input that is not documented as valid. In addition I wanted to test the ability to send malformed data as well, so I adjusted the entity object structure.